tests: Share SpawnCmdWithLogger between coverage and no coverage scenarios
