+++ tests/framework/e2e/etcd_spawn.go
@@ -15,6 +15,7 @@
 package e2e
 
 import (
+	"os"
 	"strings"
 
 	"go.uber.org/zap"
@@ -23,7 +24,8 @@ import (
 )
 
 var (
-	initBinPath func(string) binPath
+	initBinPath    func(string) binPath
+	additionalArgs func() ([]string, error)
 )
 
 func SpawnCmd(args []string, envVars map[string]string) (*expect.ExpectProcess, error) {
@@ -32,4 +34,23 @@ func SpawnCmd(args []string, envVars map[string]string) (*expect.ExpectProcess, 
 
 func SpawnNamedCmd(processName string, args []string, envVars map[string]string) (*expect.ExpectProcess, error) {
 	return SpawnCmdWithLogger(zap.NewNop(), args, envVars, processName)
+}
+
+func SpawnCmdWithLogger(lg *zap.Logger, args []string, envVars map[string]string, name string) (*expect.ExpectProcess, error) {
+	wd, err := os.Getwd()
+	if err != nil {
+		return nil, err
+	}
+
+	newArgs, err := additionalArgs()
+	if err != nil {
+		return nil, err
+	}
+	env := mergeEnvVariables(envVars)
+	lg.Info("spawning process",
+		zap.Strings("args", args),
+		zap.String("working-dir", wd),
+		zap.String("name", name),
+		zap.Strings("environment-variables", env))
+	return expect.NewExpectWithEnv(args[0], append(args[1:], newArgs...), env, name)
 }
