+++ tests/framework/e2e/etcd_spawn_nocov.go
@@ -17,18 +17,11 @@
 
 package e2e
 
-import (
-	"os"
-
-	"go.uber.org/zap"
-
-	"go.etcd.io/etcd/pkg/v3/expect"
-)
-
 const noOutputLineCount = 0 // regular binaries emit no extra lines
 
 func init() {
 	initBinPath = initBinPathNoCov
+	additionalArgs = additionalArgsNoCov
 }
 
 func initBinPathNoCov(binDir string) binPath {
@@ -40,16 +33,6 @@ func initBinPathNoCov(binDir string) binPath {
 	}
 }
 
-func SpawnCmdWithLogger(lg *zap.Logger, args []string, envVars map[string]string, name string) (*expect.ExpectProcess, error) {
-	wd, err := os.Getwd()
-	if err != nil {
-		return nil, err
-	}
-	env := mergeEnvVariables(envVars)
-	lg.Info("spawning process",
-		zap.Strings("args", args),
-		zap.String("working-dir", wd),
-		zap.String("name", name),
-		zap.Strings("environment-variables", env))
-	return expect.NewExpectWithEnv(args[0], args[1:], env, name)
+func additionalArgsNoCov() ([]string, error) {
+	return []string{}, nil
 }
